Refactor UpdateKnot, CreateKnot, DeleteKnot
